Add "carbonphp/carbon-doctrine-types" package and update other dependencies
This commit includes the addition of "carbonphp/carbon-doctrine-types" to the composer.lock file, which enables usage of Carbon in Doctrine. Moreover, existing dependencies such as "myclabs/deep-copy", "nesbot/carbon", "phpmailer/phpmailer" and several others have been updated to their latest versions. This contributes to maintaining the project up-to-date with the latest improvements and security fixes.